+++ pages/_app.js
@@ -3,6 +3,7 @@ import { useRouter } from 'next/router';
 import Head from 'next/head';
 import { DefaultSeo } from 'next-seo';
 import * as gtag from '../lib/gtag';
+import '../styles/global.css';
 import 'tailwindcss/tailwind.css';
 
 function MyApp({ Component, pageProps }) {
@@ -24,11 +25,6 @@ function MyApp({ Component, pageProps }) {
     <div className="text-lg bg-imsetyWhite text-imsetyBlack dark:bg-imsetyBlack dark:text-imsetyWhite">
       <Head>
         <link rel="icon" href="/favicon.ico" />
-        <link rel="preconnect" href="https://fonts.gstatic.com" />
-        <link
-          href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap"
-          rel="stylesheet"
-        />
       </Head>
       <DefaultSeo
         title="Imsety Taylor"