+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+### Added
+
+- Cool scale animation when you hover a project.
+
+### Changed
+
+- Made project section wider.
+- Rounded off some things.
+- Swapped "Hey!" with "Greetings!"
+- Changed front from Poppins to [Hubot Sans](https://github.com/github/hubot-sans).
+- Centered hero images.
+
 ## [1.6.0] - 2022-07-30
 
 ### Added